+++ config/services.yaml
@@ -6,7 +6,6 @@ services:
     Torr\Storyblok\:
         resource: ../src/*
         exclude:
-             - ../src/Data
              - ../src/Exception
-             - ../src/Structure/Field/Option
+             - ../src/Field
              - ../src/TorrStoryblokBundle.php
