+++ src/Field/Definition/RichTextField.php
@@ -1,11 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Torr\Storyblok\Structure\Field;
+namespace Torr\Storyblok\Field\Definition;
 
-use Torr\Storyblok\Data\FieldType;
-use Torr\Storyblok\Data\RichTextStyling;
+use Torr\Storyblok\Component\ComponentDefinitionInterface;
 use Torr\Storyblok\Exception\InvalidFieldConfigurationException;
-use Torr\Storyblok\Structure\Component\ComponentTypeDefinition;
+use Torr\Storyblok\Field\FieldType;
+use Torr\Storyblok\Field\RichText\RichTextStyling;
 use Torr\Storyblok\Validator\DataValidator;
 
 final class RichTextField extends AbstractField
@@ -13,10 +13,10 @@ final class RichTextField extends AbstractField
 	/**
 	 * @inheritDoc
 	 *
-	 * @param array<class-string<ComponentTypeDefinition>> $filterComponents
-	 * @param array<string>                                $filterComponentGroups
-	 * @param array<RichTextStyling>                       $toolbarOptions
-	 * @param array<array<string, string>>                 $styleOptions
+	 * @param array<class-string<ComponentDefinitionInterface>> $filterComponents
+	 * @param array<string>                                     $filterComponentGroups
+	 * @param array<RichTextStyling>                            $toolbarOptions
+	 * @param array<array<string, string>>                      $styleOptions
 	 */
 	public function __construct (
 		string $label,
