+++ src/Field/Definition/TableField.php
@@ -1,18 +1,18 @@
 <?php declare(strict_types=1);
 
-namespace Torr\Storyblok\Structure\Field;
+namespace Torr\Storyblok\Field\Definition;
 
-use Torr\Storyblok\Data\FieldType;
+use Torr\Storyblok\Field\FieldType;
 use Torr\Storyblok\Validator\DataValidator;
 
-final class BooleanField extends AbstractField
+final class TableField extends AbstractField
 {
 	/**
 	 * @inheritDoc
 	 */
 	protected function getInternalStoryblokType () : FieldType
 	{
-		return FieldType::Boolean;
+		return FieldType::Table;
 	}
 
 	/**
