+++ src/Field/Option/LanguagesSource.php
@@ -1,13 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Torr\Storyblok\Structure\Field\Option;
+namespace Torr\Storyblok\Field\Option;
 
-final class RemoteJsonSource implements ChoiceSourceInterface
+final class LanguagesSource implements ChoiceSourceInterface
 {
 	/**
 	 */
 	public function __construct (
-		private readonly string $url,
 		private readonly bool $showEmptyOption = true,
 	) {}
 
@@ -17,10 +16,8 @@ final class RemoteJsonSource implements ChoiceSourceInterface
 	public function toManagementApiData () : array
 	{
 		return [
-			"source" => "external",
+			"source" => "internal_languages",
 			"exclude_empty_option" => !$this->showEmptyOption,
-			"external_datasource" => $this->url,
 		];
 	}
-
 }
