+++ src/Field/RichText/RichTextStyling.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Torr\Storyblok\Data;
+namespace Torr\Storyblok\Field\RichText;
 
 enum RichTextStyling : string
 {
